Added usernames, public realms end-points + some bugfix and cleanups.

+++ agate-core/src/main/java/org/obiba/agate/service/RealmConfigService.java
@@ -1,7 +1,9 @@
 package org.obiba.agate.service;
 
+import org.obiba.agate.domain.Group;
 import org.obiba.agate.domain.RealmConfig;
 import org.obiba.agate.domain.RealmStatus;
+import org.obiba.agate.domain.User;
 import org.obiba.agate.repository.RealmConfigRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Component;
@@ -12,6 +14,7 @@ import javax.inject.Inject;
 import javax.validation.constraints.NotNull;
 import java.util.Collection;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Component
 @Transactional
@@ -22,13 +25,23 @@ public class RealmConfigService {
 
   private final RealmConfigRepository realmConfigRepository;
 
+  private final UserService userService;
+
+  private final GroupService groupService;
+
   @Inject
-  public RealmConfigService(RealmConfigRepository realmConfigRepository) {
+  public RealmConfigService(RealmConfigRepository realmConfigRepository,
+                            UserService userService,
+                            GroupService groupService) {
     this.realmConfigRepository = realmConfigRepository;
+    this.userService = userService;
+    this.groupService = groupService;
   }
 
   public RealmConfig save(@NotNull RealmConfig config) {
     assert config != null;
+
+    groupService.ensureGroupsByName(config.getGroups());
     RealmConfig saved = config;
 
     if (saved.isNew()) {
@@ -59,6 +72,12 @@ public class RealmConfigService {
 
   public void delete(@NotNull String name) {
     Assert.notNull(name, "Realm config name cannot be null.");
+
+    List<String> usernames = getUsernames(name);
+    if (usernames.size() > 0) {
+      throw NotOrphanRealmException.withName(name);
+    }
+
     RealmConfig config = findConfig(name);
     if (config != null) realmConfigRepository.delete(config);
   }
@@ -80,9 +99,13 @@ public class RealmConfigService {
 
   public void updateGroups(@NotNull String name, @NotNull Collection<String> groups) {
     Assert.notNull(name, "Realm config name cannot be null.");
+    groupService.ensureGroupsByName(groups);
     RealmConfig config = getConfig(name);
     config.setGroups(groups);
     realmConfigRepository.save(config);
   }
 
+  public List<String> getUsernames(String name) {
+    return userService.findUsers().stream().map(User::getRealm).filter(name::equals).collect(Collectors.toList());
+  }
 }
